# ingester: reduce default -server.grpc-max-concurrent-streams to 500