--- a/operations/helm/tests/graphite-enabled-values-generated/mimir-distributed/templates/ingester/ingester-statefulset.yaml
+++ b/operations/helm/tests/graphite-enabled-values-generated/mimir-distributed/templates/ingester/ingester-statefulset.yaml
@@ -93,6 +93,7 @@ spec:
             - "-config.expand-env=true"
             - "-config.file=/etc/mimir/mimir.yaml"
             - "-ingester.ring.instance-availability-zone=zone-default"
+            - "-server.grpc-max-concurrent-streams=500"
           volumeMounts:
             - name: config
               mountPath: /etc/mimir