--- a/operations/helm/tests/metamonitoring-values-generated/mimir-distributed/templates/mimir-config.yaml
+++ b/operations/helm/tests/metamonitoring-values-generated/mimir-distributed/templates/mimir-config.yaml
@@ -96,7 +96,6 @@ data:
     runtime_config:
       file: /var/mimir/runtime.yaml
     server:
-      grpc_server_max_concurrent_streams: 1000
       grpc_server_max_connection_age: 2m
       grpc_server_max_connection_age_grace: 5m
       grpc_server_max_connection_idle: 1m