--- a/operations/mimir/ingester.libsonnet
+++ b/operations/mimir/ingester.libsonnet
@@ -15,14 +15,12 @@
     {
       target: 'ingester',
 
+      // Server config.
       'server.http-listen-port': $._config.server_http_port,
 
       // Ring config.
       'ingester.ring.num-tokens': 512,
       'ingester.ring.unregister-on-shutdown': $._config.unregister_ingesters_on_shutdown,
-
-      // Limits config.
-      'server.grpc-max-concurrent-streams': 10000,
 
       // Blocks storage.
       'blocks-storage.tsdb.dir': '/data/tsdb',