--- a/operations/mimir/ingester.libsonnet
+++ b/operations/mimir/ingester.libsonnet
@@ -15,7 +15,6 @@
     {
       target: 'ingester',
 
-      // Server config.
       'server.http-listen-port': $._config.server_http_port,
 
       // Ring config.