… operations/helm/tests/test-enterprise-legacy-label-values-generated/mimir-distributed/templates/mimir-config.yaml
… operations/helm/tests/test-enterprise-values-generated/mimir-distributed/templates/mimir-config.yaml
… operations/mimir-tests/test-all-components-generated.yaml
… operations/mimir-tests/test-all-components-with-tsdb-head-early-compaction-generated.yaml
… operations/mimir-tests/test-all-components-without-chunk-streaming-generated.yaml
… operations/mimir-tests/test-autoscaling-custom-target-utilization-generated.yaml
… operations/mimir-tests/test-autoscaling-generated.yaml
… operations/mimir-tests/test-consul-generated.yaml
… operations/mimir-tests/test-consul-multi-zone-generated.yaml
… operations/mimir-tests/test-consul-ruler-disabled-generated.yaml
… operations/mimir-tests/test-continuous-test-generated.yaml
… operations/mimir-tests/test-defaults-generated.yaml
… operations/mimir-tests/test-deployment-mode-migration-generated.yaml
… operations/mimir-tests/test-env-vars-generated.yaml
… operations/mimir-tests/test-etcd-replicas-generated.yaml
… operations/mimir-tests/test-extra-runtime-config-generated.yaml
… operations/mimir-tests/test-helm-parity-generated.yaml
… operations/mimir-tests/test-memberlist-cluster-label-migration-step-0-before-generated.yaml
… operations/mimir-tests/test-memberlist-cluster-label-migration-step-1-generated.yaml
… operations/mimir-tests/test-memberlist-cluster-label-migration-step-2-generated.yaml
… operations/mimir-tests/test-memberlist-cluster-label-migration-step-3-generated.yaml
… operations/mimir-tests/test-memberlist-migration-step-0-before-generated.yaml
… operations/mimir-tests/test-memberlist-migration-step-1-generated.yaml
… operations/mimir-tests/test-memberlist-migration-step-2-generated.yaml
… operations/mimir-tests/test-memberlist-migration-step-3-generated.yaml
… operations/mimir-tests/test-memberlist-migration-step-4-generated.yaml
… operations/mimir-tests/test-memberlist-migration-step-5-generated.yaml
… operations/mimir-tests/test-memberlist-migration-step-6-final-generated.yaml
… operations/mimir-tests/test-memcached-mtls-generated.yaml
… operations/mimir-tests/test-multi-zone-generated.yaml