@dimitarvdimitrov 2023-10-02
this is again ready for a review. I changed the value from 100 to 500, since that's what we estimated as an upper bound to the number of gRPC streams that an ingester might have to handle from distributors.

We've been running this for a few weeks at grafana labs and don't see any problems related to latency, so it should be safe to make this change for everyone.